wechat/requirements.txt
# 微信公众号订阅模块依赖

# HTTP 请求
requests>=2.31.0
httpx>=0.25.0

# 配置管理
pyyaml>=6.0.1
python-dotenv>=1.0.0

# RSS 解析
feedparser>=6.0.10

# 数据库
sqlalchemy>=2.0.0

# AI 客户端
litellm>=1.30.0

# 邮件发送
# (使用标准库 smtplib，无需额外依赖)

# HTML 模板
jinja2>=3.1.0

# Markdown 转 HTML
markdown>=3.5.0

# 日期时间处理
python-dateutil>=2.8.2
pytz>=2024.1

# 定时任务
schedule>=1.2.0

# HTML 处理
beautifulsoup4>=4.12.0
lxml>=5.0.0
